using validator where possible, splitting thinx.js router away
