+++ lib/thinx/deployment.js
@@ -18,23 +18,8 @@ var debug_deployment = app_config.debug.deployment || false;
 
 module.exports = class Deployment {
 
-	validatedOwner(owner) {
-		// TODO: Maximum length 64
-		var regex = /[!@#$%&*()_\-+={[}\]|\:;"'<,>.?\/\\~`]/g;
-		return owner.replace(regex, "");
-	}
-
-	validatedUDID(udid) {
-		// TODO: Maximum length 64
-		if ((typeof(udid) === "undefined") || (udid === null)) {
-			console.log("invalid udid in ", this.arguments.callee);
-		}
-		var regex = /[!@#$%&*()_\+={[}\]|\:;"'<,>.?\/\\~`]/g;
-		return udid.replace(regex, "");
-	}
-
 	deployPathForOwner(owner) {
-		let v_owner = this.validatedOwner(owner);
+		let v_owner = Validator.owner(owner);
 		var user_path = app_config.data_root + app_config.deploy_root + "/" + v_owner;
 		return user_path;
 	}
