+++ lib/thinx/device.js
@@ -41,12 +41,6 @@ module.exports = class Device {
 		this.client = redis.createClient(Globals.redis_options());
 	}
 
-
-	validatedUDID(udid) {
-		var regex = /[!@#$%&*()_\+={[}\]|\:;"'<,>.?\/\\~`]/g;
-		return udid.replace(regex, "");
-	}
-
 	// private
 	_storeOTT(body, callback) {
 		if (typeof(body) === "undefined") {
@@ -783,11 +777,6 @@ module.exports = class Device {
 		}); // get
 	}
 
-  validatedOwner(owner) {
-    var regex = /[!@#$%&*()_\-+={[}\]|\:;"'<,>.?\/\\~`]/g;
-    return owner.replace(regex, "");
-  }
-
 	register(body, api_key, websocket, callback, req) {
 
 		var reg = body;
@@ -852,7 +841,7 @@ module.exports = class Device {
 
 		var push = reg.push;
 		var alias = reg.alias;
-		var owner = this.validatedOwner(reg.owner);
+		var owner = this.Validator.owner(reg.owner);
 		var version = reg.version;
 
 		var timezone_offset = 0;
